Made final changes to the login page, oauth should work now
Keep in mind, that the functionality f the login api call is changed. I briefly described this in Teams. Also, keep in mind that your browser will most likely cache the account you last logged in as, so when you go to the login page you will be automatically logged in unless you go incognito